# frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "PDF Reader",
        "description": "Local PDF RAG Bot with Chat, Mindmap, and Podcast features",
        "version": "1.0.0"
    },
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication",
                    "Authentication"
                ],
                "summary": "Register",
                "description": "Register a new user",
                "operationId": "register_auth_register_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserRegisterRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserRegisterResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication",
                    "Authentication"
                ],
                "summary": "Login",
                "description": "Login user and create session",
                "operationId": "login_auth_login_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserLoginResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication",
                    "Authentication"
                ],
                "summary": "Logout",
                "description": "Logout user and invalidate session",
                "operationId": "logout_auth_logout_post",
                "parameters": [
                    {
                        "name": "session_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Session Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout-all": {
            "post": {
                "tags": [
                    "Authentication",
                    "Authentication"
                ],
                "summary": "Logout All",
                "description": "Logout from all devices (invalidate all user sessions)",
                "operationId": "logout_all_auth_logout_all_post",
                "parameters": [
                    {
                        "name": "session_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Session Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/profile": {
            "get": {
                "tags": [
                    "Authentication",
                    "Authentication"
                ],
                "summary": "Get Profile",
                "description": "Get user profile information",
                "operationId": "get_profile_auth_profile_get",
                "parameters": [
                    {
                        "name": "session_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Session Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserProfileResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/change-password": {
            "post": {
                "tags": [
                    "Authentication",
                    "Authentication"
                ],
                "summary": "Change Password",
                "description": "Change user password",
                "operationId": "change_password_auth_change_password_post",
                "parameters": [
                    {
                        "name": "session_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Session Id"
                        }
                    },
                    {
                        "name": "old_password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Old Password"
                        }
                    },
                    {
                        "name": "new_password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "New Password"
                        }
                    },
                    {
                        "name": "confirm_password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Confirm Password"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "UserRegisterRequest": {
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "title": "Email"
                    },
                    "username": {
                        "type": "string",
                        "maxLength": 50,
                        "minLength": 3,
                        "title": "Username"
                    },
                    "password": {
                        "type": "string",
                        "maxLength": 128,
                        "minLength": 8,
                        "title": "Password"
                    },
                    "confirm_password": {
                        "type": "string",
                        "maxLength": 128,
                        "minLength": 8,
                        "title": "Confirm Password"
                    }
                },
                "type": "object",
                "required": [
                    "email",
                    "username",
                    "password",
                    "confirm_password"
                ],
                "title": "UserRegisterRequest",
                "description": "User registration request model"
            },
            "UserRegisterResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "message": {
                        "type": "string",
                        "title": "Message"
                    },
                    "user_id": {
                        "type": "string",
                        "title": "User Id"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "username": {
                        "type": "string",
                        "title": "Username"
                    }
                },
                "type": "object",
                "required": [
                    "status",
                    "message",
                    "user_id",
                    "email",
                    "username"
                ],
                "title": "UserRegisterResponse",
                "description": "User registration response model"
            },
...
        }
    }
}